add the calculation for fiber distribution and the method is using area percentage value

+++ src/test/testprog.py
@@ -1,6 +1,6 @@
-#!/usr/bin env python
+#! /usr/bin/env python
 
-# Copyright (C) 2009  B. Malengier
+# Copyright (C) 2010  P. Li, B. Malengier
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -17,16 +17,20 @@
 # Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 """
-    Main program that reads in ini file for fiber simulation and decides how 
-    to handle it
+    Main program that reads in ini file for yarn 2D simulation and decides 
+    how to handle it
 """
 #-------------------------------------------------------------------------
 #
 # Python modules
 #
 #-------------------------------------------------------------------------
+from __future__ import division
 import sys, os, shutil
 import getopt
+import time
+import numpy as N
+import scipy as S
 
 #-------------------------------------------------------------------------
 #
@@ -35,7 +39,7 @@ import getopt
 #-------------------------------------------------------------------------
 from lib.utils.utils import set_outputdir
 import const
-import fiber1d.config as conf
+import test.config as conf
 
 #-------------------------------------------------------------------------
 #
@@ -62,41 +66,55 @@ def main(argv=None):
         print "Error parsing the arguments: %s " % argv[1:]
         sys.exit(0)
     if leftargs:
-        print 'fiber.py does not understand argument %s' % leftargs
+        print 'yarn2d.py does not understand argument %s' % leftargs
         sys.exit(0)
 
     inifile = conf.INIFILE_DEFAULT
     outputdir = const.DATA_DIR
+    onlymesh = False
     for opt_ix in range(len(options)):
         option, value = options[opt_ix]
         if option in ( '-i', '--inifile'):
             inifile = value
         elif option in ('-o', '--outputdir'):
             outputdir = value
+        elif option in ('--mesh'):
+            onlymesh = True
     
     #Parse ini file to obtain parameters.
-    cfg = conf.Fiber1dConfigManager.get_instance(inifile)
+    print inifile
+    cfg = conf.TestFiberConfigManager.get_instance(inifile)
+    if not hasattr(cfg, 'onlymesh'):
+        cfg.onlymesh = onlymesh
     
     #create outputdir if not existing
     if not os.path.isdir(outputdir):
         os.mkdir(outputdir)
     #create outputdir for this run, remove if existing
     outputdir = outputdir + os.sep + os.path.basename(inifile)
-    if not os.path.isdir(outputdir):
-        os.mkdir(outputdir)
+    #determine whether we make a new file for yarn2d or read the old file; the
+    #default value is 'False', which means the new file will be generated when
+    #the programme runs
+    read_old_file = cfg.get('general.read')
+    if not read_old_file:
+        if not os.path.isdir(outputdir):
+            os.mkdir(outputdir)
+        else:
+            shutil.rmtree(outputdir)
+            os.mkdir(outputdir)
+        set_outputdir(outputdir)
+        #store the ini file in the outputdir so the experiment can be repeated
+        shutil.copy(inifile, outputdir)
     else:
-        shutil.rmtree(outputdir)
-        os.mkdir(outputdir)
-    set_outputdir(outputdir)
-    #store the ini file in the outputdir so the experiment can be repeated
-    shutil.copy(inifile, outputdir)
+        set_outputdir(outputdir)
+        print "ready to read the old file"
     
     #create the correct model, and run it
-    from fiber1d.fibermodel import FiberModel
-    model = FiberModel(cfg)
-    
-    #pass further execution to the model
-    model.run_step(step)#(wait=True, output=True)
+    from test.testfiber import TestFiber
+    model = TestFiber(cfg)
+
+    #pass further execution to the mode
+    model.run()
 
 if __name__ == "__main__":
     main()
